requirements.txt
Flask
python-dotenv
openai
#sqlite